+++ packages/payload/packages/xns/plugins/record/src/helpers/validation/factory/spec/tsconfig.json
