+++ packages/payload/packages/xns/plugins/record/src/helpers/validation/index.ts
@@ -1,2 +1,2 @@
-export * from './private/index.ts'
-export * from './public/index.ts'
+export * from './factory/index.ts'
+export * from './validators/index.ts'
