+++ packages/payload/packages/xns/plugins/record/src/helpers/validation/validators/spec/tsconfig.json
